# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple manga reader app with rate, favorite and comment features",
        "title": "Manga Explorer API",
        "contact": {
            "name": "arcorium",
            "url": "github.com/arcorium",
            "email": "arcorium.l@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/mangas": {
            "get": {
                "description": "Get random manga with limit query",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga"
                ],
                "summary": "Random Manga",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "total response manga",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/dto.MinimalMangaResponse"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "create new manga",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga"
                ],
                "summary": "Create Manga",
                "parameters": [
                    {
                        "description": "manga create input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.MangaCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.SuccessWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.SuccessResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "data": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.ErrorWrapper"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.ErrorResponse"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "details": {
                                                            "type": "object"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.ErrorWrapper": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/dto.ErrorResponse"
                }
            }
        },
...
        "dto.GenreResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.MangaCreateInput": {
            "type": "object",
            "required": [
                "desc",
                "genres",
                "origin",
                "publication_year",
                "status",
                "title"
            ],
            "properties": {
                "desc": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "origin": {
                    "type": "string"
                },
                "publication_year": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.MinimalMangaResponse": {
            "type": "object",
            "properties": {
                "cover_url": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GenreResponse"
                    }
                },
                "id": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "rate": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "total_comment": {
                    "type": "integer"
                },
                "total_rater": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ResponsePage": {
            "type": "object",
            "properties": {
                "elements": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total_elements": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "page": {
                    "$ref": "#/definitions/dto.ResponsePage"
                }
            }
        },
        "dto.SuccessWrapper": {
            "type": "object",
            "properties": {
                "success": {
                    "$ref": "#/definitions/dto.SuccessResponse"
                }
            }
        },
...
    }
}